backend/requirements.txt
django==4.2.19
djangorestframework==3.15.2
django-dotenv==1.4.2
psycopg2==2.9.10
requests==2.31.0
google-api-python-client==2.114.0
python-dotenv==1.0.1
drf-spectacular==0.27.1
gunicorn==21.2.0
coreapi==2.3.3
django-cors-headers==4.3.1
django-filter==23.5
django-redis==5.4.0
whitenoise==6.6.0
sentry-sdk==1.31.0